Changed admin interface
Creating and deleting not allowd when database does not exist

+++ admin/admin_home.php
@@ -129,40 +129,51 @@ function deleteDatabase($conn)
 function createUser($conn)
 {
     global $database_hostname;
-    // Check if the user exists before attempting to drop it
-    $checkUserQuery = "SELECT 1 FROM mysql.user WHERE user = '" . DATABASE_USERNAME . "' AND host = '$database_hostname'";
-    $checkUserResult = $conn->query($checkUserQuery);
 
-    if ($checkUserResult->num_rows > 0) {
-        // User already exists
-        echo "User already exist.<br>";
+    // Check if the database exists
+    $query = "SELECT SCHEMA_NAME FROM INFORMATION_SCHEMA.SCHEMATA WHERE SCHEMA_NAME = '" . DATABASE_NAME . "'";
+    $result = $conn->query($query);
+
+    if ($result->num_rows > 0) {
+        // Database exists, create user
+        // Check if the user exists before attempting to drop it
+        $checkUserQuery = "SELECT 1 FROM mysql.user WHERE user = '" . DATABASE_USERNAME . "' AND host = '$database_hostname'";
+        $checkUserResult = $conn->query($checkUserQuery);
+
+        if ($checkUserResult->num_rows > 0) {
+            // User already exists
+            echo "User already exist.<br>";
+        } else {
+
+            $createUserQuery = "CREATE USER '" . DATABASE_USERNAME . "'@'$database_hostname' IDENTIFIED BY '" . DATABASE_PASSWORD . "'";
+            $createUserResult = $conn->query($createUserQuery);
+            if ($createUserResult === TRUE) {
+                echo "User created successfully.<br>";
+            } else {
+                echo "Error creating user: " . $conn->error . "<br>";
+            }
+
+            // Grant privileges
+            $grantPrivilegesQuery = "GRANT SELECT, INSERT, UPDATE, DELETE ON " . DATABASE_NAME . ".* TO '" . DATABASE_USERNAME . "'@'$database_hostname'";
+            $grantPrivilegesResult = $conn->query($grantPrivilegesQuery);
+            if ($grantPrivilegesResult === TRUE) {
+                echo "Privileges granted successfully.<br>";
+            } else {
+                echo "Error granting privileges: " . $conn->error . "<br>";
+            }
+
+            // Flush privileges
+            $flushPrivilegesQuery = "FLUSH PRIVILEGES";
+            $flushPrivilegesResult = $conn->query($flushPrivilegesQuery);
+            if ($flushPrivilegesResult === TRUE) {
+                echo "privileges flushed successfully.<br>";
+            } else {
+                echo "Error flushing privileges: " . $conn->error . "<br>";
+            }
+        }
     } else {
-
-        $createUserQuery = "CREATE USER '" . DATABASE_USERNAME . "'@'$database_hostname' IDENTIFIED BY '" . DATABASE_PASSWORD . "'";
-        $createUserResult = $conn->query($createUserQuery);
-        if ($createUserResult === TRUE) {
-            echo "User created successfully.<br>";
-        } else {
-            echo "Error creating user: " . $conn->error . "<br>";
-        }
-
-        // Grant privileges
-        $grantPrivilegesQuery = "GRANT SELECT, INSERT, UPDATE, DELETE ON " . DATABASE_NAME . ".* TO '" . DATABASE_USERNAME . "'@'$database_hostname'";
-        $grantPrivilegesResult = $conn->query($grantPrivilegesQuery);
-        if ($grantPrivilegesResult === TRUE) {
-            echo "Privileges granted successfully.<br>";
-        } else {
-            echo "Error granting privileges: " . $conn->error . "<br>";
-        }
-
-        // Flush privileges
-        $flushPrivilegesQuery = "FLUSH PRIVILEGES";
-        $flushPrivilegesResult = $conn->query($flushPrivilegesQuery);
-        if ($flushPrivilegesResult === TRUE) {
-            echo "privileges flushed successfully.<br>";
-        } else {
-            echo "Error flushing privileges: " . $conn->error . "<br>";
-        }
+        // Database does not exist
+        echo "Database does not exist, cannot create user without database.<br>";
     }
 }
 
@@ -170,22 +181,54 @@ function createUser($conn)
 function deleteUser($conn)
 {
     global $database_hostname;
-    // Check if the user exists before attempting to drop it
-    $checkUserQuery = "SELECT 1 FROM mysql.user WHERE user = '" . DATABASE_USERNAME . "' AND host = '$database_hostname'";
-    $checkUserResult = $conn->query($checkUserQuery);
 
-    if ($checkUserResult->num_rows > 0) {
-        // User exists, drop it
-        $deleteUserQuery = "DROP USER '" . DATABASE_USERNAME . "'@'$database_hostname'";
-        $deleteUserResult = $conn->query($deleteUserQuery);
-        if ($deleteUserResult === TRUE) {
-            echo "User deleted successfully.<br>";
+    // Check if the database exists
+    $query = "SELECT SCHEMA_NAME FROM INFORMATION_SCHEMA.SCHEMATA WHERE SCHEMA_NAME = '" . DATABASE_NAME . "'";
+    $result = $conn->query($query);
+
+    if ($result->num_rows > 0) {
+        // Database exists, create user
+
+        // Check if the user exists before attempting to drop it
+        $checkUserQuery = "SELECT 1 FROM mysql.user WHERE user = '" . DATABASE_USERNAME . "' AND host = '$database_hostname'";
+        $checkUserResult = $conn->query($checkUserQuery);
+
+        if ($checkUserResult->num_rows > 0) {
+            // User already exists
+            
+            // Grant privileges
+            $grantPrivilegesQuery = "REVOKE ALL PRIVILEGES ON *.* FROM '" . DATABASE_USERNAME . "'@'$database_hostname';";
+            $grantPrivilegesResult = $conn->query($grantPrivilegesQuery);
+            if ($grantPrivilegesResult === TRUE) {
+                echo "Privileges revoked successfully.<br>";
+            } else {
+                echo "Error granting privileges: " . $conn->error . "<br>";
+            }
+
+            // Flush privileges
+            $flushPrivilegesQuery = "FLUSH PRIVILEGES";
+            $flushPrivilegesResult = $conn->query($flushPrivilegesQuery);
+            if ($flushPrivilegesResult === TRUE) {
+                echo "privileges flushed successfully.<br>";
+            } else {
+                echo "Error flushing privileges: " . $conn->error . "<br>";
+            }
+
+            // delete user
+            $deleteUserQuery = "DROP USER '" . DATABASE_USERNAME . "'@'$database_hostname'";
+            $deleteUserResult = $conn->query($deleteUserQuery);
+            if ($deleteUserResult === TRUE) {
+                echo "User deleted successfully.<br>";
+            } else {
+                echo "Error deleting user: " . $conn->error;
+            }
         } else {
-            echo "Error deleting user: " . $conn->error;
+            // User does not exist
+            echo "User does not exist.<br>";
         }
     } else {
-        // User does not exist
-        echo "User does not exist.<br>";
+        // Database does not exist
+        echo "Database does not exist, cannot create user without database.<br>";
     }
 }
 
@@ -193,56 +236,76 @@ function deleteUser($conn)
 // Function to create the table
 function createTable($conn)
 {
-    // Select database
-    $conn->select_db(DATABASE_NAME);
+    // Check if the database exists
+    $query = "SELECT SCHEMA_NAME FROM INFORMATION_SCHEMA.SCHEMATA WHERE SCHEMA_NAME = '" . DATABASE_NAME . "'";
+    $result = $conn->query($query);
 
-    // Check if table exists before attempting to creat it
-    $checkTableQuery = "SHOW TABLES LIKE '" . TABLE_NAME . "'";
-    $checkTableResult = $conn->query($checkTableQuery);
+    if ($result->num_rows > 0) {
+        // Database exists
+        // Select database
+        $conn->select_db(DATABASE_NAME);
 
-    if ($checkTableResult->num_rows > 0) {
-        // Table already exists
-        echo "Table already exist<br>";
-    } else {
-        // Table does not exist, create it
-        $createTableQuary = "CREATE TABLE " . TABLE_NAME . " (
+        // Check if table exists before attempting to creat it
+        $checkTableQuery = "SHOW TABLES LIKE '" . TABLE_NAME . "'";
+        $checkTableResult = $conn->query($checkTableQuery);
+
+        if ($checkTableResult->num_rows > 0) {
+            // Table already exists
+            echo "Table already exist<br>";
+        } else {
+            // Table does not exist, create it
+            $createTableQuary = "CREATE TABLE " . TABLE_NAME . " (
             user_id INT PRIMARY KEY AUTO_INCREMENT,
             email VARCHAR(50) NOT NULL UNIQUE,
             username VARCHAR(20) NOT NULL UNIQUE,
             password_hash VARCHAR(60) NOT NULL
         )";
 
-        $createTableResult = $conn->query($createTableQuary);
-        if ($createTableResult === TRUE) {
-            echo "Table created successfully<br>";
-        } else {
-            echo "Error creating table: " . $conn->error . "<br>";
+            $createTableResult = $conn->query($createTableQuary);
+            if ($createTableResult === TRUE) {
+                echo "Table created successfully<br>";
+            } else {
+                echo "Error creating table: " . $conn->error . "<br>";
+            }
         }
+    } else {
+        // Database does not exists
+        echo "Database does not exists, cannot create table without database.";
     }
 }
 
 // Function to delete the table
 function deleteTable($conn)
 {
-    // Select database
-    $conn->select_db(DATABASE_NAME);
+    // Check if the database exists
+    $query = "SELECT SCHEMA_NAME FROM INFORMATION_SCHEMA.SCHEMATA WHERE SCHEMA_NAME = '" . DATABASE_NAME . "'";
+    $result = $conn->query($query);
 
-    // Check if table exists before attempting to delete it
-    $checkTableQuery = "SHOW TABLES LIKE '" . TABLE_NAME . "'";
-    $checkTableResult = $conn->query($checkTableQuery);
+    if ($result->num_rows > 0) {
+        // Database exists, delete it
+        // Select database
+        $conn->select_db(DATABASE_NAME);
 
-    if ($checkTableResult->num_rows > 0) {
-        // Table exists, delete it
-        $dropTableQuary = "DROP TABLE " . TABLE_NAME;
-        $dropTableResult = $conn->query($dropTableQuary);
-        if ($dropTableResult === TRUE) {
-            echo "Table dropped successfully.<br>";
+        // Check if table exists before attempting to delete it
+        $checkTableQuery = "SHOW TABLES LIKE '" . TABLE_NAME . "'";
+        $checkTableResult = $conn->query($checkTableQuery);
+
+        if ($checkTableResult->num_rows > 0) {
+            // Table exists, delete it
+            $dropTableQuary = "DROP TABLE " . TABLE_NAME;
+            $dropTableResult = $conn->query($dropTableQuary);
+            if ($dropTableResult === TRUE) {
+                echo "Table dropped successfully.<br>";
+            } else {
+                // Table does not exist
+                echo "Error dropping table: " . $conn->error . "<br>";
+            }
         } else {
-            // Table does not exist
-            echo "Error dropping table: " . $conn->error . "<br>";
+            echo "Table does not exist.<br>";
         }
     } else {
-        echo "Table does not exist.<br>";
+        // Database does not exist
+        echo "Database does not exists, cannot delete table without database.";
     }
 }
